requirements.txt
boto3>=1.35.0
botocore>=1.35.0
streamlit>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sounddevice>=0.5.0
numpy>=1.26.0
scipy>=1.14.0
